Allow to perform actions on issues
When creating a new issue, we show a popup with certain actions to
perform on the newly created issue.

This adds that ~same popup that allows to:
* Copy ID
* Copy branch name
* Open in browser

The popup is added when settings an issue to the context directly or by
choosing one of the user's issues, as well as adds a new command to show
the current issue in the context, and show a popup for it, so that the
user can always perform certain actions on the issue they're currently
working on.